requirements.txt
playwright==1.44.0
pytest==7.3.1
python-dotenv==1.0.1
